requirements.txt
click==7.1.2
pyaml==19.12.0
PyYAML==5.3
selenium==3.141.0
easygui==0.98.1
